Unify signature of scalar [de]compression routines
Compression: input uint16_t, output uint32_t
Decompression: input uint32_t, output uint16_t.

+++ cbmc/proofs/scalar_compress_q_16/scalar_compress_q_16_harness.c
@@ -22,7 +22,7 @@
  *
  */
 void harness(void) {
-  int32_t u;
+  uint16_t u;
 
   /* Contracts for this function are in poly.h */
   uint32_t d = scalar_compress_q_16(u);
